Introduce GKScaleMode: Add support for an additional scaling mode: GKScaleModeAspectFill.
GKScaleModeAspectFit is the old (and default) type, which would scale the entire image to fit into
the cropping region, even if that means that there'd be empty space on the left/right or top/bottom.

GKScaleModeAspectFill is the new type (opt-in), which scales up the entire image so that there's never
any empty space in the cropping region.

The changes introduced here should be backwards compatible and require no code changes to maintain the old behavior.