GKImagePicker: Work around an issue that could cause the statusBar to disappear.
UIImagePickerController (when used to take camera pictures) hides the
status bar. Given that the class inherits from UINavigationController,
one might assume that one can simply use -pushViewController and
friends. Unfortunately there seems to be some side-effect in the class'
implementation that will cause the status bar not to be made visible
again if the entire VC gets dismissed.
We work around that by calling -[setStatusBarHidden: NO] prior to
pushing, which is somewhat nasty.